cleanup

+++ history.php
@@ -1,6 +1,16 @@
 <?php
 	function delOld(){
-		$db = mysqli_connect("localhost","datalogger","datalogger") or die("DB Connect error");
+		$servername = "localhost";
+		$username = "datalogger";
+		$password = "datalogger";
+		$dbname = "datalogger";
+
+		// Create connection
+		$db = mysqli_connect($servername, $username, $password, $dbname);
+		// Check connection
+		if (!$db) {
+			die("Connection failed: " . mysqli_connect_error());
+		}
 		mysqli_select_db("datalogger");
 		$q="delete from datalogger";
 		mysqli_query($db, $q);
@@ -9,7 +19,17 @@
 	}
 
 	function hist($sensor){
-		$db = mysqli_connect("localhost","datalogger","datalogger") or die("DB Connect error");
+		$servername = "localhost";
+		$username = "datalogger";
+		$password = "datalogger";
+		$dbname = "datalogger";
+
+		// Create connection
+		$db = mysqli_connect($servername, $username, $password, $dbname);
+		// Check connection
+		if (!$db) {
+			die("Connection failed: " . mysqli_connect_error());
+		}
 		mysqli_select_db("datalogger");
 		for ($i=0;$i<=23;$i++)
 		{
